Update buffer management and improve naming.
The key change in this commit is the update in the implementation of buffer management in Structures.kt. Instead of using an array, buffers now leverage pointers for improved control. In addition, a transformer is added in RenderPassDescriptor.kt and a check condition is built in helpers.kt to avoid initialization if the data is empty. Lastly, the window title in GlfwMain.kt has been updated for accuracy.