Implement ImageBitmapHolder and update scene list
Implemented the ImageBitmapHolder by using the BufferedImage class. The width and height no longer throw a "Not yet implemented" exception, and instead provide the actual values from the BufferedImage object. In addition, the scenes list in Application.kt has been updated, and the texture field in Structures.kt is now nullable.