Add support for Xlib window system in wgpu4k
This update includes the creation of a new class `WGPUXlibWindowSurfaceDescriptor` in `Structures.kt` which helps in communicating with the Xlib window system. Necessary changes have been made in `Functions.kt` and `WGPU.kt` to initialize instances of this new class. Additionally, the `GlfwMain.kt` example code has also been modified accordingly to work with the updated API.