Add compute pass capabilities and buffer writing method
This commit introduces a compute pass capability in the CommandEncoder class and adds a method for writing buffers in the Queue class. It further defines the additional compute pass internals, where functionalities including setting pipelines and dispatching workgroups were indicated. A start for creation of compute pipeline in Device class is also laid out.