Add 'webgpu4kotlin' image and adjust Particles scene
The new 'webgpu4kotlin' image has been added to the asset manager for both the web-js and glfw examples. In the Particles scene, this image has been used instead of 'Di3d'. Various adjustments have also been made to the scene, including a changed deltaTime, altered view translations and rotations, and the removal of a duplicate view declaration. The Particles scene has also been added to 'availableScenes' in Application.kt.