Update WGPUVertexState structure and introduce new CCodeGenerator class
This commit modifies the WGPUVertexState class in the Structures.kt file, specifically it changes the constantCount variable type and replaces "constants" with "constantsPtr" in the FieldOrder attribute. A new class - CCodeGenerator - is introduced into the codebase with functionality to debug Structure data with indentation. Additionally, 'CCodeGeneratorKtTest' is created for testing the 'log' functionality, and minor changes are made in build.gradle.kts.