Update texture loading and remove BufferedImage dependencies
This commit updates the method for texture loading to use a more flexible Bitmap32 format instead of relying on BufferedImage. The conversion methods for BufferedImage were removed due to their redundancy after this change. Now, all images are read as BMP32 before being converted to the necessary format for texture mapping.