Update library versions and modify test configuration
This commit updates the versions of used libraries in the libs.versions.toml file and adds 'download' and 'kotest-runner-junit5' to the list. Furthermore, it modifies the test configuration in the test.yml and build.gradle.kts files, changing the Gradle command from 'test' to 'check', and updating the implementation from 'org.opentest4j' to 'libs.kotest.runner.junit5'. Also, it sets up more detailed logging for test failures and passes.