Refactor Buffer.jvm.kt and remove useless methods in Functions.kt
The refactor involves changing the way the 'Buffer' class interacts with 'WebGPU_h' library in Buffer.jvm.kt. This promotes code reusability and makes the code more organized.
In addition, the commit removes numerous functions in Functions.kt that are no longer needed. This can improve readability and maintainability within the project.