Refactor buffer mapping and cleanup unused functions
In this commit, we have made a significant refactor to the buffer mapping in 'Buffer.jvm.kt' and 'BindGroupDescriptor.kt'. Instead of previously used 'handler2', we directly use 'WGPUBufferImpl(resource.buffer.handler.toPointer())' to handle buffer. Furthermore, we have cleaned up several unused functions in the 'Functions.kt' file to improve simplicity and readability of the code.