(instable) Update RenderPipelineDescriptor class and related mappings in wgpu
The code updates the RenderPipelineDescriptor class in the wgpu package to use val instead of var. The adjustments for this change are then propagated through related mappers and functions within the mapper package, primarily replacing old mappings for RenderPipelineDescriptor and its subclasses with updated versions. This also includes changes in how RenderPipelineDescriptor is used elsewhere in the wgpu package.