Refactor code to handle enumeration conversions
The commit restructures the handling of enumeration conversions by adding a new function, toFlagInt(), in the EnumerationWithValue interface. All individual enumeration classes' or() functions are removed, and the toFlagInt() function is shifted to the EnumerationWithValue interface for more unified and streamlined operation. Additionally, the commit includes renaming of WGPUBufferDescriptor.kt to BufferDescriptor.kt, along with respective changes in the subclasses. New test cases have also been added to verify the new approach.