Refactored BindGroupDescriptor class and mapping functions
Refactored the BindGroupDescriptor class, making its properties immutable. Also, modified how the mapping of BindGroupDescriptor is accomplished, using Arena.map function instead of the mapper class. This aims to make the code cleaner and easier to understand. Old unused imports and classes, such as WGPUBindGroupLayoutImpl and BindGroupLayoutTransformer, have also been removed as part of this refactor.