Refactor bloom functionality to use ES modules
The bloom functionality was refactored to use ES modules, with the relevant script in spookyball.js updated. Cleanup was done on unused bloom.js in the webgpu directory and the references to bloom.js were replaced with references to spookyball.js. Additionally, the bloom.kt file was made exportable for JavaScript with @JsExport annotation added.