Add default positioning and scaling to Transform class and Vector3D conversions
Added default position and scale variables to the Transform class and ensured their use in this class. To facilitate the handling of this data in JavaScript, Float32Array conversion functionality was added to the Vector3D class. This makes it possible for Kotlin and JavaScript to work cohesively with 3D vector data.