Add vec3 transform features to spookyball
This update modifies the Spookyball example to add a vec3 transform function for 3D vector calculations in WebGL. The added function transforms a 3D point using a 4x4 matrix. Changes in the code also include replacing some matrix calculations with a new MyVector3 implementation and a function to insert values into the Float32Array.