Add camera and general uniforms buffer
This commit introduces two new buffer groups - a camera bind group cluster and a general uniforms bind group cluster in the skinned mesh class. Uniforms are used for the camera and general aspects of the visual representation in the graphics. This allows for further development and functionality in the application's graphics representation techniques.