Update various GLTF render and mesh details
This commit updates several details in the GLTF rendering and mesh-related code. Specifically, it comments out the skinBindGroupLayout in the GLTF2Render and SkinnedMesh files, adjusts the format of the print statement in the GLTF2Render, modifies the vertex attribute location indicies and shader generation in RotatingMeshCube, and changes the asset path in AssetManager. These changes are meant to adjust the rendering and potentially improve performance.