Refactor and optimize mesh rendering in RotatingMeshCube
The commit includes optimizations to the rendering of meshes in the RotatingMeshCube class. It provides a more efficient method of handling vertex and index buffers and adjusts the render loop to iterate over all mesh primitives and their attributes. A late initialization of the GLTF2RenderContext also been introduced to the Application's initialize function.