Refactor GLB import logic to use VfsFile instead of ArrayBuffer
This commit replaces the use of ArrayBuffer in the uploadGLBModel method with VfsFile for improved file handling. Now, GLB files are read directly into memory and converting it into Int8Array instead of reading it as ArrayBuffer. Additionally, the GLTFBufferView has been updated to use BufferView from GLTF2 and remove redundant byteOffset and byteStride calculations.