Move `webgpu-gltf2` to `examples` and delete unnecessary files
The main action performed in this commit was the transfer of `webgpu-gltf2` related files to the `examples` directory. Along with the files transfer, several unnecessary files were deleted, such as `Camera.kt`, `main.kt`, and files related to the build process and rendering. Additionally, relevant changes were made to the `Application.kt` and `SkinnedMesh.kt` files in the `examples` directory. The `AssetManager.kt` file was also updated accordingly.