Remove redundant compose main classes and refactor scene classes
Deleted files "ComposeMain.kt" and "GlfwMain.kt" as they are not needed anymore. Several scene classes were also refactored. The primary change involved removing calls to surface.present() within the render() functions of scene files, as it is no longer necessary since it will be handled by the framework. Dependency injection was also added to some scene classes to allow passing in of WGPUContext and GenericAssetManager.