Refactor rendering process and code clean-up
The commit implements the copyTextureToBuffer function in CommandEncoder.js and clears numerous warnings related to headless rendering. This includes removing AssetManager references from scene classes and making various updates in jsMain and commonMain files. Additionally, a clean-up was performed in the Application.kt file to further support the rendering process. This results in a more streamlined and efficient rendering process and a cleaner, warning-free codebase.