Update screenshot path fetching and enhance rendering flow
The fetching of the screenshot path in the render.js.kt has been updated, it no longer throws an error when the path is not found; it simply assigns null. In the endToEndWebserver.kt file, a flag has been introduced to check when rendering has finished. This allows for more controlled screenshot capturing, waiting for the rendering to end before taking a screenshot. Additionally, unnecessary print statements from render.kt were removed for cleaner output.