Update screenshot bundling in workflow
Modifications were made to the .github/workflows/test.yml file to enhance how screenshots are handled. Now, screenshots gathered from the chromium and jvm workflows will be bundled according to the runner's operating system. This change will reflect in the naming of the output zip files, improving organization and searchability.