Add maven-publish to kotlin plugin and configure publishing
In the build.gradle.kts file, a maven-publish plugin for kotlin has been added, which allows publishing artifacts to Maven repositories. A condition has also been configured that sets a different publishing repository depending on whether the version is 'SNAPSHOT' or not. Authentication for the GitLab repository was also included.