Update GitHub workflows and build scripts for versioning
The GitHub workflows and build scripts have been updated to handle versioning in a better way. The logic for managing SNAPSHOT versioning has moved to a new isSnapshot() method. This provides clearer, centralized control over versioning. Now, the version to publish can be passed as an optional input to the workflow_dispatch event, and the SNAPSHOT variable can be set to true or false as needed.