Refactor render pipeline descriptor mapping to use createJsObject function
The code has been refactored to use the `createJsObject` function in the existing mapping process of the `RenderPipelineDescriptor`. This change was implemented to replace the previously used explicit object declaration across methods for converting input configurations to GPU compatible configurations. Moreover, the same change is reflected in the `createRenderPipeline` method calls where the mapping process is involved.