Update PipelineLayout mapping and debug shader creation
Updated the 'createPipelineLayout' function to use the new mapping function in the Arena object. Also, a 'println' command has been added in the 'ShaderCache.kt' and 'domain.kt' for the purpose of debugging the shader creation process, which involves generating new shaders and creating the render pipeline. Enabled the 'SkinnedMeshScene' in the 'Scene.kt'.