Remove unnecessary native GPU handler classes
This commit removes many unnecessary native GPU handler classes, which were not utilised and cluttering the codebase. This also resulted in various functionalities being overridden with no meaningful implementation, causing confusion. With the removal of these handler classes, the codebase becomes cleaner, allowing for a more comprehensive understanding of the project structure and functionality.