Update JVM target and enhance dependencies settings
This commit updates the JVM target to 11 in the wgpu4k module. The setting of the JVM target has been moved from the tasks section to the platform-specific configuration, providing a more accurate control. In addition, the dependencies of certain libraries in the common module have been enhanced from implementation to API. New properties are also added to the Gradle properties file for better compatibility. Finally, .gitignore has been updated to exclude local.properties files.