Add iOS support to the multiplatform project
This commit adds iOS support for the multiplatform project. A new build.gradle.kts file is added for the iOS task, and the reference to it is included only for Mac OS X in settings.gradle.kts. The main.kt file was also created under the iOS source directory. Changes are also made in the 'wgpu4k' directory to support the iOS building process.