Add 'nothing' function to iOS and simulator examples
Renamed main.iosArm64.kt and main.iosSimulatorArm64.kt to main.kt in both iOS and simulator projects. These changes include the addition of a new function, 'nothing', and the subsequent invocation in AAPLViewController.m. This is part of a broader effort to streamline file names and simplify function calls across all examples.